+++ src/interface/src/app/services/plan.service.ts
@@ -1,23 +1,16 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { map, Observable, take } from 'rxjs';
-import { CreatePlanPayload, Plan, PreviewPlan } from '@types';
+import {
+  CreatePlanPayload,
+  Creator,
+  Pagination,
+  Plan,
+  PreviewPlan,
+} from '@types';
 import { GeoJSON } from 'geojson';
 import { environment } from '../../environments/environment';
 import { Params } from '@angular/router';
-
-interface PlanResults {
-  count: number;
-  next?: string;
-  previous?: string;
-  results: PreviewPlan[];
-}
-
-export interface Creator {
-  id: number;
-  email: string;
-  full_name: string;
-}
 
 @Injectable({
   providedIn: 'root',
@@ -95,16 +88,16 @@ export class PlanService {
     });
   }
 
-  getPlanPreviews(params: Params): Observable<PlanResults> {
+  getPlanPreviews(params: Params) {
     let url = environment.backend_endpoint.concat(this.v2basePath);
 
-    return this.http.get<PlanResults>(url, {
+    return this.http.get<Pagination<PreviewPlan>>(url, {
       withCredentials: true,
       params: params,
     });
   }
 
-  listPlanCreators() {
+  getCreators() {
     let url = environment.backend_endpoint.concat('/planning/v2/creators/');
     return this.http.get<Creator[]>(url, {
       withCredentials: true,
