+++ src/interface/src/app/standalone/planning-areas/planning-areas.component.scss
@@ -147,11 +147,3 @@ table.planning-areas {
   align-items: center;
   justify-content: center;
 }
-
-// placeholders, remove once the components are ready.
-.filter-placeholder {
-  border: 1px solid $color-text-gray;
-  padding: 8px;
-  border-radius: 4px;
-}
-
